描画システムの変更、Fontクラス追加

+++ include/EZGS/Color.hpp
@@ -2,20 +2,21 @@
  * @file Color.hpp
  * @brief êFä÷òA
  * @author Yoshito Nakaue
- * @date 2020/08/20
+ * @date 2020/08/21
  */
 
 
 #pragma once
+#include <cstdint>
 
 namespace ezgs
 {
     struct Color
     {
     public:
-        float red;
-        float green;
-        float blue;
-        float alpha;
+        uint8_t red;
+        uint8_t green;
+        uint8_t blue;
+        uint8_t alpha;
     };
 }
